requirements.txt
pillow
requests
zhdate